Fixed init stop to wait for the process to stop
The "sleep 5" in the restart action must have seemed like a clever hack
at the time, but unfortunately the dpkg triggers will run a "stop"
followed by a "start", potentially in a very short space of time. Java
(with its lack of proper signal handling) takes a second or so to shut
down, so the typical effect of an upgrade is for the daemon to attempt
to start before the RMI registry port is available. This apparently is
not a serious enough problem for the daemon to fail to start, rather it
just soldiers on, with empty result sets returned for queries that need
RMI on the affected host.

Just using start-stop-daemon with --retry fixes the problem. Confirmed
on my lucene test VM.

Change-Id: I52f5973fd161f0564a09a78094b8067bee78e283