Bug: 46295 Fix error parsing InitialiseSettings.php
Current code in PHPParser.getContentNamespaces() expects namespace values to be
integers and gets an exception when it encounters a symbolic value such as NS_MAIN.
This patch adds support for the 16 core namespaces by hardcoding a table of those
symbolic values and their numeric equivalents. This seems like a better alternative
than parsing Defines.php especially since a comment in that file strongly forbids
changing this mapping.

Bug: 46295
Change-Id: Iad2ed7b2a236f6073d86ef3b39e54fe394032c21