Dump entire global config to a file.
This patch dumps the entire singleton GlobalConfiguration object to the file
/var/tmp/global.txt. Most of the functionality of Lucene search is driven by
this singleton object so this dump will be useful while we are stabilizing
search. The file is around 10MB and the dump is controlled by the DBG variable
in GlobalConfiguration which can be set to 'false' when we deem lsearchd to
be adequately stable. I decided to make a separate file for this output since
I didn't want it to be clobbered by the rolling file appenders.

Since this dump is done just once during initialization, there should be no
performance impact thereafter.

Change-Id: I60b1be9441b3652514c76e5160750a1e2dcc4ff0